MAke all file uploads id hashed by data
