+++ src/data/utils/uid.ts
@@ -1,12 +1,34 @@
-const abc = "abcdefghijklmnopqrstuvwxyz";
-const letters = abc.concat(abc.toUpperCase());
+// @ts-ignore
+import MD5 from "md5.js";
 
-const ALLOWED_CHARS = `0123456789${letters}`;
+// DHIS2 UID :: /^[a-zA-Z][a-zA-Z0-9]{10}$/
+const asciiLetters = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ";
+const asciiNumbers = "0123456789";
+const asciiLettersAndNumbers = asciiLetters + asciiNumbers;
+const uidSize = 10;
+const range = Array.from(new Array(uidSize).keys());
+const uidStructure = [asciiLetters, ...range.map(() => asciiLettersAndNumbers)];
+const maxHashValue = uidStructure.map(cs => cs.length).reduce((acc, n) => acc * n, 1);
 
-const NUMBER_OF_CODEPOINTS = ALLOWED_CHARS.length;
-const CODESIZE = 11;
+/* Return pseudo-random UID from seed string */
+export function getUid(seed: string): string {
+    const md5hash: string = new MD5().update(seed).digest("hex");
+    const nHashChars = Math.ceil(Math.log(maxHashValue) / Math.log(16));
+    const hashInteger = parseInt(md5hash.slice(0, nHashChars), 16);
+    const result = uidStructure.reduce(
+        (acc, chars) => {
+            const { n, uid } = acc;
+            const nChars = chars.length;
+            const quotient = Math.floor(n / nChars);
+            const remainder = n % nChars;
+            const uidChar = chars[remainder];
+            return { n: quotient, uid: uid + uidChar };
+        },
+        { n: hashInteger, uid: "" }
+    );
 
-const CODE_PATTERN = /^[a-zA-Z]{1}[a-zA-Z0-9]{10}$/;
+    return result.uid;
+}
 
 function randomWithMax(max: number) {
     return Math.floor(Math.random() * max);
@@ -14,15 +36,17 @@ function randomWithMax(max: number) {
 
 export function generateUid(): string {
     // First char should be a letter
-    let randomChars = letters.charAt(randomWithMax(letters.length));
+    let randomChars = asciiLetters.charAt(randomWithMax(asciiLetters.length));
 
-    for (let i = 1; i < CODESIZE; i += 1) {
-        randomChars += ALLOWED_CHARS.charAt(randomWithMax(NUMBER_OF_CODEPOINTS));
+    for (let i = 1; i <= uidSize; i += 1) {
+        randomChars += asciiLettersAndNumbers.charAt(randomWithMax(asciiLettersAndNumbers.length));
     }
 
     return randomChars;
 }
 
+const uidRegex = /^[a-zA-Z]{1}[a-zA-Z0-9]{10}$/;
+
 export function isValidUid(code: string | undefined | null): boolean {
-    return !!code && CODE_PATTERN.test(code);
+    return !!code && uidRegex.test(code);
 }
